requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Columnar==1.3.1
cryptography==3.4.7
finnhub-python==2.4.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.3
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
td-ameritrade-python-api==0.3.5
termcolor==1.1.0
toolz==0.11.1
urllib3==1.26.4
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
